clean ups and refactorings

+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/internal/ResourceUIValidatorExtension.java
@@ -177,7 +177,9 @@ public class ResourceUIValidatorExtension extends MarkerEraser implements IResou
 
 		try {
 			for (IContainer container : containers) {
-				container.deleteMarkers(null, true, IResource.DEPTH_INFINITE);
+				if (container.exists()) {
+					container.deleteMarkers(null, true, IResource.DEPTH_INFINITE);
+				}
 			}
 		} catch (CoreException e) {
 			e.printStackTrace();
